Consolidate E2E tests and clean up unused code
- Move liquidity slippage test into AMM Functionality describe block
- Remove unused resetToCleanState function and network-reset utility
- Delete separate liquidity-slippage.spec.ts file
- Add Argos reporter configuration for screenshot support
- Remove .env.local from git tracking and add to .gitignore
