+++ .gitignore
@@ -43,6 +43,9 @@ testem.log
 .DS_Store
 Thumbs.db
 
+# Environment files
+.env.local
+
 .nx/cache
 .nx/workspace-data
 .cursor/rules/nx-rules.mdc
