--- apps/e2e/.env.local
@@ -1,6 +1,0 @@
-# Auto-generated contract addresses and deployment info for e2e tests
-# Generated on Sun Jul  6 16:54:14 CST 2025
-TOKEN_ADDRESS=0x5FbDB2315678afecb367f032d93F642f64180aa3
-AMM_POOL_ADDRESS=0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0
-DEPLOYMENT_BLOCK_NUMBER=3
-CHAIN_ID=31337
